% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{build_sql_api_url}
\alias{build_sql_api_url}
\title{Build url for sql inquiry}
\usage{
build_sql_api_url(inquiry, parameter = "")
}
\arguments{
\item{inquiry}{sql inquiry string}

\item{parameter}{additional parameter to sql inquiry, like file format or
file name}
}
\value{
encoded url of sql api call
}
\description{
Build url for sql inquiry
}

