% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{build_sql_batch_job_url}
\alias{build_sql_batch_job_url}
\title{Build url for checking batch sql inquiry}
\usage{
build_sql_batch_job_url(job_id)
}
\arguments{
\item{job_id}{the job id of batch sql inquiry, returned from request response}
}
\value{
encoded url of checking batch sql inquiry
}
\description{
Build url for checking batch sql inquiry
}

