% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{build_sql_batch_url}
\alias{build_sql_batch_url}
\title{Build url for creating batch sql inquiry}
\usage{
build_sql_batch_url()
}
\value{
encoded url of batch sql inquiry
}
\description{
Build url for creating batch sql inquiry
}
