% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{build_sql_url}
\alias{build_sql_url}
\title{Build url for sql and batch sql inquiry}
\usage{
build_sql_url(middle)
}
\arguments{
\item{middle}{the unique part of each individual API call}
}
\value{
sql API url
}
\description{
Build url for sql and batch sql inquiry
}
