% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{build_sync_url}
\alias{build_sync_url}
\title{Build url for checking sync status or force sync}
\usage{
build_sync_url(table_id)
}
\arguments{
\item{table_id}{the sync table id from the sync request response}
}
\value{
API url
}
\description{
The url is same for two commands, except GET or PUT
}
