% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{build_url}
\alias{build_url}
\title{Build url for import and sync API call}
\usage{
build_url(middle)
}
\arguments{
\item{middle}{the unique part of each individual API call}
}
\value{
import and sync API url
}
\description{
The import and sync API url are very similar. This function capture the
shared parts. The individual API call will add its special part as parameter
to build its url.
}
\details{
Called by \code{\link{local_import}}, \code{\link{url_common}}
}
