% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{carto_setup}
\alias{carto_setup}
\title{Read Carto user name and API key from environment}
\usage{
carto_setup()
}
\value{
a named vector holding user name and API key
}
\description{
The design goal is 1. hide details and varibles from user, only expose update
function,2. put help details in update function, 3. report error in all
functions, 4. report user name in success reading, only in package loading
and after update.
}
\details{
That means carto_setup only report error, and a separate function to provide
message string of reporting success or failure. Package loading and updating
will message this string. That function need to suppress carto_setup message
and use its own copy because the return result is a message string, not
printing message immediately.
}

