% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_carto_setup}
\alias{check_carto_setup}
\title{Message string of user name found in env}
\usage{
check_carto_setup()
}
\value{
message string
}
\description{
If user name is not found, error message is reported from all the
functions. If user name is found, this only need to be reported in package
loading or after updating the environment, not in normal function usage.
}

