% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{check_sync_status}
\alias{check_sync_status}
\title{Check sync status}
\usage{
check_sync_status(table_id)
}
\arguments{
\item{table_id}{the sync table id from the sync request response}
}
\value{
request response in JSON format
}
\description{
Sometimes there are empty link or unfinished sync reported as success in
request status. The table of all sync files actually provids more information
about status than this function.
}
\details{
Calling function without assigning return value will print response status
and response content in console. Assigning return value will print response
status only.
}
