% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{force_sync_without_time_check}
\alias{force_sync_without_time_check}
\title{Force sync a sync table}
\usage{
force_sync_without_time_check(table_id)
}
\arguments{
\item{table_id}{the sync table id from the sync request response}
}
\value{
request response in JSON format
}
\description{
There is a 15 mins wait limit since last sync. Depend on last sync time the
force sync may fail. When force sync is possible there should be a "sync now"
button in data set view page.
}

