% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_response}
\alias{get_response}
\title{Print or return the API call response}
\usage{
get_response(res, print_only = FALSE)
}
\arguments{
\item{res}{response object}

\item{print_only}{If TRUE, only print the status, no returning of content. Default FALSE.}
}
\value{
If \code{print_only} is FALSE, the response object in prettified json format
}
\description{
There are two parts in response: response status and response content. The status report whether the request is sucess, and the content have more information about the request may be useful later.
}
\details{
Run function without assigning return value will print both status and response in console. Assigning return value will only print the status in console, then save the content in prettified json format.
}

