% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{list_sync_tables_dt}
\alias{list_sync_tables_dt}
\title{Get the sync files table and convert to data.table}
\usage{
list_sync_tables_dt()
}
\value{
A data.table holding sync file table
}
\description{
If calling function without assigning return value, both request status and
reponse will be printed in console. If assigning a return value to variable,
only status will be printed.
}
\details{
Note the time in response is UTC time. To convert it into local time, use
\code{with_tz(ymd_hms(time))}
}

