% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{local_import}
\alias{local_import}
\title{Upload local file to Carto}
\usage{
local_import(file_name)
}
\arguments{
\item{file_name}{the local file path}
}
\value{
request response
}
\description{
Calling function without assigning return value will print response status
and response content in console. Assigning return value will print response
status only.
}
