% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{remove_sync}
\alias{remove_sync}
\title{Remove sync for remote file}
\usage{
remove_sync(table_id)
}
\arguments{
\item{table_id}{the sync table id from the sync request response}
}
\value{
request response in JSON format
}
\description{
Just remove the sync schedule, not the existing table in Carto.
}
\details{
Finished sync tables should appear in Carto.com data set page. Empty or wrong
sync tables will not appear and cannot be delete from web UI.
}
