% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{setup_key}
\alias{setup_key}
\title{Set up Carto user name and API key with environment}
\usage{
setup_key()
}
\value{
setup status message
}
\description{
All functions need a Carto user name and API key.
}
\details{
\enumerate{
  \item run \code{file.edit("~/.Renviron")} to edit the environment
variable file
  \item add two lines
    \itemize{ \item \code{carto_acc = "your
user name"}
    \item \code{carto_api_key = "your api key"}
    }
  \item run \code{setup_key()}.
 }

Note if you want to remove the key and deleted the lines from
\code{~/.Renviron}, the key could still exist in environment. Restart R
session to make sure it was removed.

For adding key or changing key value, edit and run \code{setup_key()} is
enough.
}

