% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql_batch_cancel}
\alias{sql_batch_cancel}
\title{Cancel a Batch sql inquiry}
\usage{
sql_batch_cancel(job_id)
}
\arguments{
\item{job_id}{job id of previously submitted Batch sql inquiry}
}
\value{
request status
}
\description{
Cancel a Batch sql inquiry
}
