% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql_inquiry_df}
\alias{sql_inquiry_df}
\title{Run sql inquiry and save result in data frame}
\usage{
sql_inquiry_df(inquiry)
}
\arguments{
\item{inquiry}{sql inquiry string}
}
\value{
result in data frame
}
\description{
Calling function without assigning return value will print response status
and response content in console. Assigning return value will print response
status only.
}

