% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql_inquiry_dt}
\alias{sql_inquiry_dt}
\title{Run sql inquiry and save result in data.table}
\usage{
sql_inquiry_dt(inquiry)
}
\arguments{
\item{inquiry}{sql inquiry string}
}
\value{
result in data.table
}
\description{
Calling function without assigning return value will print response status
and response content in console. Assigning return value will print response
status only.
}

