% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql_inquiry_save_geojson}
\alias{sql_inquiry_save_geojson}
\title{Run sql inquiry and save result file as GeoJSON}
\usage{
sql_inquiry_save_geojson(inquiry, filepath)
}
\arguments{
\item{inquiry}{sql inquiry string}

\item{filepath}{output file path}
}
\value{
request status
}
\description{
A wrapper to \code{sql_inquiry_save(inquiry, filepath, "&format=GeoJSON")}
}
\examples{
sql_inquiry_save("SELECT * FROM table_1 limit 2", "e:/test.json")
}

