% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{test_connection}
\alias{test_connection}
\title{Simple sql inquiry used for testing setup}
\usage{
test_connection()
}
\value{
request response in JSON format
}
\description{
After setting Carto user name and API key, this function can be used to test
if the setup is successful.
}
\details{
It's just a wrapper to \code{sql_inquiry("SELECT 1")}
}
