% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{update_env}
\alias{update_env}
\title{Update environment variables}
\usage{
update_env()
}
\value{
setup status message
}
\description{
\code{readRenviron("~/.Renviron")} then check if the setup is correct
}

