% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{url_common}
\alias{url_common}
\title{The commong part of url to call import or sync API with a remote link}
\usage{
url_common(link, quoted_guessing = FALSE, api)
}
\arguments{
\item{link}{remote file link}

\item{quoted_guessing}{If TRUE, Carto will also guess quoted columns. Default
FALSE.}

\item{api}{import or sync API name}
}
\value{
request response
}
\description{
Used \code{\link{build_url}}. Called by \code{\link{url_import}},
\code{\link{url_sync}} Note the API call to list sync table used different
format
}

