% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{url_import}
\alias{url_import}
\title{Import a remote file link into Carto}
\usage{
url_import(link, quoted_guessing = FALSE)
}
\arguments{
\item{link}{remote file link}

\item{quoted_guessing}{If TRUE, Carto will also guess quoted columns. Default
FALSE.}
}
\value{
request response
}
\description{
Import a remote file link into Carto
}

