% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{url_sync}
\alias{url_sync}
\title{Sync with a remote file link}
\usage{
url_sync(link, quoted_guessing = FALSE)
}
\arguments{
\item{link}{remote file link}

\item{quoted_guessing}{If TRUE, Carto will also guess quoted columns. Default
FALSE.}
}
\value{
request response
}
\description{
Sync with a remote file link
}

