% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{carto_setup}
\alias{carto_setup}
\title{Setup Carto user name and API key}
\usage{
carto_setup()
}
\value{
a named vector holding user name and API key
}
\description{
All functions need a Carto user name and API key.
}
\details{
1. run \code{file.edit("~/.Renviron")} and add \code{Sys.setenv(carto_acc =
"your user name")}; \code{Sys.setenv(carto_api_key = "your api key")}

2. run \code{readRenviron("~/.Renviron")} or \code{update_api_key()} to
update environment variables.

Check package readme for more details.
}

