% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{carto_setup}
\alias{carto_setup}
\title{Setup Carto user name and API key}
\usage{
carto_setup()
}
\value{
a named vector holding user name and API key
}
\description{
All functions need a Carto user name and API key.
}
\details{
1. To add user name and API key in console or script, run \code{carto_env <-
c(carto_acc = "your user name", carto_api_key = "your api key")}

2. To save them in R user profile so it's more permanent and not exposed in
script, run \code{file.edit("~/.Rprofile")} and add
\code{Sys.setenv(carto_acc = "your user name")};
\code{Sys.setenv(carto_api_key = "your api key")}

Check package readme for more details about two approaches.
}

