% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{check_carto_setup}
\alias{check_carto_setup}
\title{Check if carto user name and API key is available}
\usage{
check_carto_setup()
}
\value{
setup status message string
}
\description{
Check if carto user name and API key is available
}

