% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{convert_dropbox_link}
\alias{convert_dropbox_link}
\title{Convert dropbox share link into direct access file url}
\usage{
convert_dropbox_link(link = readClipboard())
}
\arguments{
\item{link}{the dropbox share link.}
}
\value{
dropbox direct access file url
}
\description{
Will take clipboard content if parameter not provided (windows only).
}
