% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_response}
\alias{get_response}
\title{Print or return the API call response}
\usage{
get_response(res, content_echo = TRUE)
}
\arguments{
\item{res}{response object}

\item{content_echo}{whether to print request content in console}
}
\value{
If \code{no_echo} is FALSE, the response object in prettified json
  format
}
\description{
There are two parts in response: response status and response content. The
status report whether the request is success, and the content is the return
value of API call.
}
\details{
Run function without assigning return value will print both status and
response in console. Assigning return value will only print the status in
console, then save the content in prettified json format.

Function will stop for http error.
}

