% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{list_sync_tables}
\alias{list_sync_tables}
\title{List all tables in sync}
\usage{
list_sync_tables()
}
\value{
request response in JSON format
}
\description{
If calling function without assigning return value, both request status and
reponse will be printed in console. If assigning a return value to variable,
only status will be printed.
}
\details{
Note the time in response is UTC time
}

