% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{list_sync_tables_dt}
\alias{list_sync_tables_dt}
\title{Get the sync files table and convert to data.table}
\usage{
list_sync_tables_dt()
}
\value{
A data.table holding sync file table
}
\description{
Calling function without assigning return value will print response status
and response content in console. Assigning return value will print response
status only.
}
\details{
Note the time in response is UTC time. To convert it into local time, use
\code{lubridate::with_tz(lubridate::ymd_hms(time))}
}

