% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{sql_inquiry}
\alias{sql_inquiry}
\title{Run sql inquiry}
\usage{
sql_inquiry(inquiry, parameter = "")
}
\arguments{
\item{inquiry}{sql inquiry string}

\item{parameter}{additional parameter to sql inquiry, like file format or
file name}
}
\value{
request response in JSON format
}
\description{
Calling function without assigning return value will print response status
and response content in console. Assigning return value will print response
status only.
}

