% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{url_sync}
\alias{url_sync}
\title{Sync with a remote file by url}
\usage{
url_sync(link, quoted_guessing = FALSE)
}
\arguments{
\item{link}{remote file link}

\item{quoted_guessing}{If TRUE, Carto will also guess quoted columns. Default
FALSE.}
}
\value{
request response
}
\description{
Calling function without assigning return value will print response status
and response content in console. Assigning return value will print response
status only.
}

